+++ packages/actions/src/index.js
@@ -2,7 +2,7 @@ export { prefixedValueMirror as makeActionTypes } from '@redux-tools/utils';
 
 export { default as configureActionCreator } from './configureActionCreator';
 export { default as makePayloadMetaActionCreator } from './makePayloadMetaActionCreator';
-export { default as makeConstantActionCreator } from './makeConstantActionCreator';
+export { default as makeEmptyActionCreator } from './makeEmptyActionCreator';
 export { default as makePayloadActionCreator } from './makePayloadActionCreator';
 
 export { default as isErrorAction } from './isErrorAction';